src/main/java/com/devcomanda/easymeetup/controller/MeetupController.java
package com.devcomanda.easymeetup.controller;

import com.devcomanda.easymeetup.entity.Meetup;
import com.devcomanda.easymeetup.service.MeetupService;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;
import org.springframework.web.bind.annotation.RestController;

@RestController
@RequestMapping("/meetup")
public class MeetupController {

    private final MeetupService meetUpService;

    public MeetupController(final MeetupService meetUpService) {
        this.meetUpService = meetUpService;
    }

    @RequestMapping(value = "/save", method = RequestMethod.POST)
    public Meetup saveMeetup(@RequestBody Meetup meetup){
        meetUpService.saveOrUpdateMeetup(meetup);
        return meetup;
    }
}
